+++ src/FullHealth/Configuration/HealMode.cs
@@ -1,0 +1,7 @@
+namespace FullHealth;
+
+public enum HealMode : byte
+{
+    All = 0,
+    Self = 1
+}