+++ src/FullHealth/Configuration/WorkMode.cs
@@ -1,0 +1,7 @@
+namespace FullHealth;
+
+public enum WorkMode : byte
+{
+    Host = 0,
+    HostAndClient = 1
+}