+++ src/FullHealth/Patches/FullHealthMultiPlayerPatch.cs
@@ -22,32 +22,9 @@ public class FullHealthMultiPlayerPatch
             return;
         }
 
-        if (!SemiFunc.IsMasterClient())
+        if (Configuration.WorkMode.Value == WorkMode.Host && !SemiFunc.IsMasterClient())
         {
             Plugin.Logger.LogDebug("This is not a master client");
-            return;
-        }
-
-        Task.Run(HealHelper.Heal);
-    }
-}
-
-[HarmonyPatch(typeof(RoundDirector), nameof(RoundDirector.StartRound))]
-public class FullHealthSinglePlayerPatch
-{
-    // ReSharper disable once InconsistentNaming
-    private static void Postfix()
-    {
-        Plugin.Logger.LogDebug("StartRound patch");
-
-        if (!ValidateHelper.IsValid())
-        {
-            return;
-        }
-
-        if (SemiFunc.IsMultiplayer())
-        {
-            Plugin.Logger.LogDebug("This is not a single player");
             return;
         }